Move fast curve graphics to new mod
